requirements.txt
pandas
biopython
rdkit
ncbi-blast+
